+++ core/Server.hpp
@@ -15,7 +15,7 @@ public:
 	Listen_Addr						getAddress() const;
 	std::string            			getName() const;
 	std::string             		getErrorPage(int code) const;
-	int                				getMaxBodySize() const;
+	ssize_t                			getMaxBodySize() const;
 	bool               				bodySizeAllowed(int bytes) const;
 	std::vector<std::string>		getMethods() const;
 	std::string						getRoot() const;
